+++ playwright.config.ts
@@ -31,7 +31,14 @@ export default defineConfig({
   projects: [
     {
       name: 'chromium',
-      use: { ...devices['Desktop Chrome'] },
+      use: {
+        ...devices['Desktop Chrome'],
+        launchOptions: {
+          args: process.env.CI
+            ? ['--disable-dev-shm-usage', '--no-sandbox', '--disable-gpu']
+            : [],
+        },
+      },
     },
     /* 
     {